feat: ✨ add unit tests for communityCreditUtil on-chain mappers
